requirements.txt
cryptography>=39.0
argon2-cffi>=21.0
pillow>=9.0
numpy>=1.24
pyqt5>=5.15
opencv-python>=4.7
pyinstaller>=5.10
scipy>=1.10  # optional: fallback edge detection / filters
